allow per-user instances of vim::plugin
The current version of this defined resource type allows one to
specify a user, but because the $title is being used as part of
the exec's command there is no way to specify the same plugin for
multiple users without getting a duplicate resource error. This
commit effectively creates a namevar `name` for this defined type,
which will allow the decoupling of the $title from the $command.

Existing behavior is still maintained.